requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
flask-login>=0.6.3
authlib>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
pymysql>=1.1.0
psycopg2-binary>=2.9.0
flask-cors>=4.0.0
playwright>=1.40.0
playwright-stealth>=2.0.0
firebase-admin>=6.5.0
google-cloud-firestore>=2.14.0
google-generativeai>=0.3.0
markdown>=3.5.0
